refactor: split on_result_select into multiple parts
